ocr/train/requirements.txt
torch
torchvision
onnx
pillow
numpy
opencv-python-headless
